+++ test_suite/tests/test_macros.rs
@@ -454,34 +454,6 @@ fn test_untagged_newtype_struct() {
 }
 
 #[test]
-fn test_adjacently_tagged_newtype_struct() {
-    #[derive(Debug, PartialEq, Serialize, Deserialize)]
-    #[serde(tag = "t", content = "c")]
-    enum E {
-        Newtype(GenericNewTypeStruct<u32>),
-        Null,
-    }
-
-    assert_de_tokens(
-        &E::Newtype(GenericNewTypeStruct(5u32)),
-        &[
-            Token::Struct { name: "E", len: 2 },
-            Token::Str("c"),
-            Token::NewtypeStruct {
-                name: "GenericNewTypeStruct",
-            },
-            Token::U32(5),
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "E",
-                variant: "Newtype",
-            },
-            Token::StructEnd,
-        ],
-    );
-}
-
-#[test]
 fn test_generic_tuple_struct() {
     assert_tokens(
         &GenericTupleStruct(5u32, 6u32),
@@ -676,400 +648,6 @@ fn test_untagged_enum() {
             Token::TupleEnd,
         ],
         "data did not match any variant of untagged enum Untagged",
-    );
-}
-
-#[test]
-fn test_adjacently_tagged_enum() {
-    #[derive(Debug, PartialEq, Serialize, Deserialize)]
-    #[serde(tag = "t", content = "c")]
-    enum AdjacentlyTagged<T> {
-        Unit,
-        Newtype(T),
-        Tuple(u8, u8),
-        Struct { f: u8 },
-    }
-
-    // unit with no content
-    assert_ser_tokens(
-        &AdjacentlyTagged::Unit::<u8>,
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 1,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // unit with no content
-    assert_de_tokens(
-        &AdjacentlyTagged::Unit::<u8>,
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // unit with tag first
-    assert_de_tokens(
-        &AdjacentlyTagged::Unit::<u8>,
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::Str("c"),
-            Token::Unit,
-            Token::StructEnd,
-        ],
-    );
-
-    // unit with content first
-    assert_de_tokens(
-        &AdjacentlyTagged::Unit::<u8>,
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("c"),
-            Token::Unit,
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // unit with excess content (f, g, h)
-    assert_de_tokens(
-        &AdjacentlyTagged::Unit::<u8>,
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("f"),
-            Token::Unit,
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::Str("g"),
-            Token::Unit,
-            Token::Str("c"),
-            Token::Unit,
-            Token::Str("h"),
-            Token::Unit,
-            Token::StructEnd,
-        ],
-    );
-
-    // newtype with tag first
-    assert_tokens(
-        &AdjacentlyTagged::Newtype::<u8>(1),
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Newtype",
-            },
-            Token::Str("c"),
-            Token::U8(1),
-            Token::StructEnd,
-        ],
-    );
-
-    // newtype with content first
-    assert_de_tokens(
-        &AdjacentlyTagged::Newtype::<u8>(1),
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("c"),
-            Token::U8(1),
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Newtype",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // optional newtype with no content field
-    assert_de_tokens(
-        &AdjacentlyTagged::Newtype::<Option<u8>>(None),
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 1,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Newtype",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // tuple with tag first
-    assert_tokens(
-        &AdjacentlyTagged::Tuple::<u8>(1, 1),
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Tuple",
-            },
-            Token::Str("c"),
-            Token::Tuple { len: 2 },
-            Token::U8(1),
-            Token::U8(1),
-            Token::TupleEnd,
-            Token::StructEnd,
-        ],
-    );
-
-    // tuple with content first
-    assert_de_tokens(
-        &AdjacentlyTagged::Tuple::<u8>(1, 1),
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("c"),
-            Token::Tuple { len: 2 },
-            Token::U8(1),
-            Token::U8(1),
-            Token::TupleEnd,
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Tuple",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // struct with tag first
-    assert_tokens(
-        &AdjacentlyTagged::Struct::<u8> { f: 1 },
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Struct",
-            },
-            Token::Str("c"),
-            Token::Struct {
-                name: "Struct",
-                len: 1,
-            },
-            Token::Str("f"),
-            Token::U8(1),
-            Token::StructEnd,
-            Token::StructEnd,
-        ],
-    );
-
-    // struct with content first
-    assert_de_tokens(
-        &AdjacentlyTagged::Struct::<u8> { f: 1 },
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("c"),
-            Token::Struct {
-                name: "Struct",
-                len: 1,
-            },
-            Token::Str("f"),
-            Token::U8(1),
-            Token::StructEnd,
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Struct",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // integer field keys
-    assert_de_tokens(
-        &AdjacentlyTagged::Newtype::<u8>(1),
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::U64(1), // content field
-            Token::U8(1),
-            Token::U64(0), // tag field
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Newtype",
-            },
-            Token::StructEnd,
-        ],
-    );
-
-    // byte-array field keys
-    assert_de_tokens(
-        &AdjacentlyTagged::Newtype::<u8>(1),
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Bytes(b"c"),
-            Token::U8(1),
-            Token::Bytes(b"t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Newtype",
-            },
-            Token::StructEnd,
-        ],
-    );
-}
-
-#[test]
-fn test_adjacently_tagged_enum_deny_unknown_fields() {
-    #[derive(Debug, PartialEq, Deserialize)]
-    #[serde(tag = "t", content = "c", deny_unknown_fields)]
-    enum AdjacentlyTagged {
-        Unit,
-    }
-
-    assert_de_tokens(
-        &AdjacentlyTagged::Unit,
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::Str("c"),
-            Token::Unit,
-            Token::StructEnd,
-        ],
-    );
-
-    assert_de_tokens_error::<AdjacentlyTagged>(
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("t"),
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::Str("c"),
-            Token::Unit,
-            Token::Str("h"),
-        ],
-        r#"invalid value: string "h", expected "t" or "c""#,
-    );
-
-    assert_de_tokens_error::<AdjacentlyTagged>(
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("h"),
-        ],
-        r#"invalid value: string "h", expected "t" or "c""#,
-    );
-
-    assert_de_tokens_error::<AdjacentlyTagged>(
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Str("c"),
-            Token::Unit,
-            Token::Str("h"),
-        ],
-        r#"invalid value: string "h", expected "t" or "c""#,
-    );
-
-    assert_de_tokens_error::<AdjacentlyTagged>(
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::U64(0), // tag field
-            Token::UnitVariant {
-                name: "AdjacentlyTagged",
-                variant: "Unit",
-            },
-            Token::U64(3),
-        ],
-        r#"invalid value: integer `3`, expected "t" or "c""#,
-    );
-
-    assert_de_tokens_error::<AdjacentlyTagged>(
-        &[
-            Token::Struct {
-                name: "AdjacentlyTagged",
-                len: 2,
-            },
-            Token::Bytes(b"c"),
-            Token::Unit,
-            Token::Bytes(b"h"),
-        ],
-        r#"invalid value: byte array, expected "t" or "c""#,
     );
 }
